Rename MobileSandbox to Avalonia.SingleProjectSandbox
