+++ dirs.proj
@@ -14,7 +14,7 @@
     <ProjectReference Remove="samples/*.Browser/*.csproj" />
     <ProjectReference Remove="samples/*.Blazor/*.csproj" />
     <ProjectReference Remove="samples/*.Tizen/*.csproj" />
-    <ProjectReference Remove="samples/MobileSandbox/*.csproj" />
+    <ProjectReference Remove="samples/SingleProjectSandbox/*.csproj" />
     <ProjectReference Remove="samples/ControlCatalog.Desktop/*.*proj" />
   </ItemGroup>
