--- samples/MobileSandbox/Platforms/Android/MainActivity.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using Android.App;
-using Android.Content.PM;
-using Android.OS;
-using Avalonia;
-using Avalonia.Android;
-
-namespace MobileSandbox;
-
-[Activity(Label = "MobileSandbox.Android", Theme = "@style/Theme.AppCompat.NoActionBar", MainLauncher = true, ConfigurationChanges = ConfigChanges.Orientation | ConfigChanges.ScreenSize)]
-public class MainActivity : AvaloniaMainActivity<App>
-{
-}
