--- samples/MobileSandbox/Platforms/Tizen/Main.cs
@@ -1,20 +1,0 @@
-using System;
-using Avalonia;
-using Avalonia.Tizen;
-using ElmSharp;
-using SkiaSharp;
-using Tizen.Applications;
-
-namespace MobileSandbox;
-
-class Program : NuiTizenApplication<App>
-{
-    protected override AppBuilder CustomizeAppBuilder(AppBuilder builder) => 
-        base.CustomizeAppBuilder(builder);
-
-    static void Main(string[] args)
-    {
-        var app = new Program();
-        app.Run(args);
-    }
-}
