--- samples/MobileSandbox/Platforms/iOS/Main.cs
@@ -1,15 +1,0 @@
-using UIKit;
-
-namespace MobileSandbox.iOS
-{
-    public class Application
-    {
-        // This is the main entry point of the application.
-        static void Main(string[] args)
-        {
-            // if you want to use a different Application Delegate class from "AppDelegate"
-            // you can specify it here.
-            UIApplication.Main(args, null, typeof(AppDelegate));
-        }
-    }
-}
