+++ samples/SingleProjectSandbox/App.axaml
@@ -1,7 +1,7 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             Name="Mobile Sandbox"
-             x:Class="MobileSandbox.App">
+             Name="Single Project Sandbox"
+             x:Class="SingleProjectSandbox.App">
   <Application.Styles>
     <FluentTheme />
   </Application.Styles>
