+++ samples/SingleProjectSandbox/MainView.axaml
@@ -1,8 +1,8 @@
-<UserControl x:Class="MobileSandbox.MainView"
+<UserControl x:Class="SingleProjectSandbox.MainView"
              xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mobileSandbox="using:MobileSandbox"
-             x:DataType="mobileSandbox:MainView">
+             xmlns:local="using:SingleProjectSandbox"
+             x:DataType="local:MainView">
   <StackPanel Margin="100 50" Spacing="50">
     <TextBlock Text="Login" Foreground="White" />
     <TextBox TextInputOptions.Multiline="True" AcceptsReturn="True" Watermark="Text" Height="200" TextWrapping="Wrap"/>
