+++ samples/SingleProjectSandbox/MainWindow.axaml
@@ -1,11 +1,10 @@
 ﻿<Window xmlns="https://github.com/avaloniaui"
-        xmlns:local="clr-namespace:MobileSandbox"
+        xmlns:local="using:SingleProjectSandbox"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         CanResize="False"
         Width="720" Height="1280"
-        Title="Mobile Sandbox"
-        Icon="/Assets/test_icon.ico"
+        Title="Single Project Sandbox"
         WindowStartupLocation="CenterScreen"
-        x:Class="MobileSandbox.MainWindow">
+        x:Class="SingleProjectSandbox.MainWindow">
   <local:MainView />
 </Window>
