+++ samples/SingleProjectSandbox/MainWindow.axaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace MobileSandbox;
+namespace SingleProjectSandbox;
 
 public partial class MainWindow : Window
 {
