+++ samples/SingleProjectSandbox/Platforms/Android/AndroidManifest.xml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android" android:installLocation="auto">
-  <application android:label="MobileSandbox.Android"></application>
+  <application android:label="SingleProjectSandbox.Android"></application>
   <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 </manifest>
