+++ samples/SingleProjectSandbox/Platforms/Browser/wwwroot/app.css
