+++ samples/SingleProjectSandbox/Platforms/Browser/wwwroot/favicon.ico
