+++ samples/SingleProjectSandbox/Platforms/Browser/wwwroot/index.html
@@ -2,7 +2,7 @@
 <html>
 
 <head>
-    <title>MobileSandbox.Browser</title>
+    <title>SingleProjectSandbox.Browser</title>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <base href="/" />
