+++ samples/SingleProjectSandbox/Platforms/Browser/wwwroot/main.js
