+++ samples/SingleProjectSandbox/Platforms/Desktop/app.manifest
