+++ samples/SingleProjectSandbox/Platforms/Tizen/shared/res/Avalonia.png
