+++ samples/SingleProjectSandbox/Platforms/iOS/Entitlements.plist
