+++ samples/SingleProjectSandbox/Platforms/iOS/Resources/LaunchScreen.xib
