+++ samples/SingleProjectSandbox/Platforms/macOS/Program.cs
@@ -1,0 +1,13 @@
+using Avalonia;
+
+namespace SingleProjectSandbox;
+
+internal static class Program
+{
+    internal static int Main(string[] args) =>
+        App.BuildAvaloniaApp()
+            .UseAvaloniaNative()
+            .UseSkia()
+            .StartWithClassicDesktopLifetime(args);
+}
+
