+++ samples/SingleProjectSandbox/SingleProjectSandbox.csproj
@@ -1,16 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>net8.0;net8.0-android;net8.0-browser;net8.0-ios</TargetFrameworks>
+    <TargetFrameworks>$(AvsCurrentTargetFramework);$(AvsCurrentAndroidTargetFramework);$(AvsCurrentBrowserTargetFramework)</TargetFrameworks>
+    <TargetFrameworks Condition="'$(IncludeIOSTargetSamples)' == 'true'">$(TargetFrameworks);$(AvsCurrentIOSTargetFramework)</TargetFrameworks>
+    <TargetFrameworks Condition="'$(IncludeMacTargetSamples)' == 'true'">$(TargetFrameworks);$(AvsCurrentMacOSTargetFramework)</TargetFrameworks>
+    <TargetFrameworks Condition="'$(IncludeTizenTargetSamples)' == 'true'">$(TargetFrameworks);$(AvsCurrentTizenTargetFramework)</TargetFrameworks>
     <OutputType>Exe</OutputType>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <Nullable>enable</Nullable>
     <IncludeAvaloniaGenerators>true</IncludeAvaloniaGenerators>
   </PropertyGroup>
-  
+
   <PropertyGroup>
     <AvaloniaSingleProject>true</AvaloniaSingleProject>
-    <ApplicationTitle>Mobile Sandbox</ApplicationTitle>
-    <ApplicationId>Avalonia.MobileSandbox</ApplicationId>
+    <ApplicationTitle>Single Project Sandbox</ApplicationTitle>
+    <ApplicationId>Avalonia.SingleProjectSandbox</ApplicationId>
   </PropertyGroup>
 
   <ItemGroup>
